Remove ngraph::Backend abstract class
