+++ test/graph_rewrites/op_by_op_capability_test.cc
@@ -29,7 +29,6 @@
 #include "logging/tf_graph_writer.h"
 #include "ngraph_bridge/default_opset.h"
 #include "ngraph_bridge/ngraph_assign_clusters.h"
-#include "ngraph_bridge/ngraph_backend.h"
 #include "ngraph_bridge/ngraph_backend_manager.h"
 #include "ngraph_bridge/ngraph_mark_for_clustering.h"
 #include "test/test_utilities.h"
