+++ test/opexecuter.h
@@ -32,7 +32,6 @@
 #include "ngraph/ngraph.hpp"
 
 #include "logging/tf_graph_writer.h"
-#include "ngraph_bridge/ngraph_backend.h"
 #include "ngraph_bridge/ngraph_backend_manager.h"
 #include "ngraph_bridge/ngraph_builder.h"
 #include "ngraph_bridge/ngraph_mark_for_clustering.h"
